# ms-order-consumer/src/main/resources/swagger.yaml
swagger: '2.0'
info:
  description: Crud Example
  version: '1.0'
  title: MS-Order - Fast-Track API Documentation
  contact:
    name: Wellington Aires Alves
    email: walves@uolinc.com
host: localhost:9999
basePath: /
tags:
  - name: order-rest
    description: Order Rest
paths:
  /orders:
    get:
      tags:
        - order-rest
      summary: list
      operationId: listUsingGET
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/OrderDto'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ExceptionResponse'

        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ExceptionResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/ExceptionResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ExceptionResponse'

      deprecated: false
    post:
      tags:
        - order-rest
      summary: create
      operationId: createUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: dto
          description: dto
          required: true
          schema:
            $ref: '#/definitions/OrderDto'
      responses:
        '200':
          description: OK
        '201':
          description: Created
          schema:
            $ref: '#/definitions/OrderDto'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ExceptionResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/ExceptionResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ExceptionResponse'

      deprecated: false
  /orders/search:
    get:
      tags:
        - order-rest
      summary: search
      operationId: searchUsingGET
      produces:
        - '*/*'
      parameters:
        - name: max_total
          in: query
          required: false
          type: string
        - name: min_total
          in: query
          required: false
          type: string
        - name: status
          in: query
          required: false
          type: string
        - name: q
          in: query
          required: false
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/OrderDto'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ExceptionResponse'
        
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/ExceptionResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ExceptionResponse'

      deprecated: false
  /orders/{id}:
    get:
      tags:
        - order-rest
      summary: findById
      operationId: findByIdUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/OrderDto'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ExceptionResponse'
        
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/ExceptionResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ExceptionResponse'

      deprecated: false
    put:
      tags:
        - order-rest
      summary: update
      operationId: updateUsingPUT
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: dto
          description: dto
          required: true
          schema:
            $ref: '#/definitions/OrderDto'
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/OrderDto'
        '201':
          description: Created
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ExceptionResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ExceptionResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/ExceptionResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ExceptionResponse'

      deprecated: false
    delete:
      tags:
        - order-rest
      summary: delete
      operationId: deleteUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/OrderDto'
        '204':
          description: No Content
          schema:
            $ref: '#/definitions/ExceptionResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ExceptionResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/ExceptionResponse'
          
      deprecated: false
definitions:
  OrderDto:
    type: object
    properties:
      description:
        type: string
      id:
        type: integer
        format: int64
      name:
        type: string
      total:
        type: number
        format: double
      status:
        type: string
        enum: [NOT_PROCESSED, PROCESSED]
    title: OrderDto
  ExceptionResponse:
    type: object
    properties:
      status_code: 
        type: integer
        format: int64
      message:
        type: string
    title: ExceptionResponse